PHP
This is the project I was telling you about. User login and registration are on the same page, Facebook-style. I learned to write the database connection in a separate file and just include it, so that's what I've always done. I'm working on the settings page right now, which you will be able to see. This is so users can customize their profile, and after I've finished this. Then I can create the index/homepage profile. It started out as a syntax reference guide, hence the name. I compared the syntax of JS, PHP, JAVA, Python, Ruby, and Swift. I wanted to structure the information in a way that was geared more towards visual learning. The languages are broken into sections. You would click a section (maybe foreach loops) and then you would see the syntax for it in a language you already know next to one you are learning. It's a simple concept, but I'm really passionate about it, and I think that it could be useful. Lol, which is all I want.